refactor: update file structure (#80)
